Moved filtering data from builders to parsers
